refactor: split docker container routes
